+++ CHANGELOG.md
@@ -29,9 +29,11 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ### Added
 - Expose `WebJira.accessAdmin()`. Resolve [JPERF-207].
 - JqlMemory modified to allow recall queries by tag/query type. Resolve [JPERF-522].
+- Add new action Search JQL a*. Resolve [JPERF-525]
 
 [JPERF-207]: https://ecosystem.atlassian.net/browse/JPERF-207
 [JPERF-522]: https://ecosystem.atlassian.net/browse/JPERF-522
+[JPERF-525]: https://ecosystem.atlassian.net/browse/JPERF-525
 
 ## [3.5.1] - 2019-05-22
 [3.5.1]: https://github.com/atlassian/jira-actions/compare/release-3.5.0...release-3.5.1